requirements.txt

python-binance
pandas
numpy
ta
PyQt5
python-telegram-bot
